+++ javaCustomSerialization2/x10.runtime/src-java/x10/core/ULong.java
@@ -27,7 +27,7 @@ final public class ULong extends Number implements StructI, java.lang.Comparable
     x10.lang.Arithmetic<ULong>, x10.lang.Bitwise<ULong>, x10.util.Ordered<ULong>
 {
     private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(ULong.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(ULong.class);
     
     public static final x10.rtt.RuntimeType<?> $RTT = Types.ULONG;
     public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
@@ -124,7 +124,7 @@ final public class ULong extends Number implements StructI, java.lang.Comparable
         serializer.write($value);
     }
 
-    public int $_get_serialization_id() {
+    public short $_get_serialization_id() {
         return _serialization_id;
     }
