+++ javaCustomSerialization2/x10.runtime/src-java/x10/core/concurrent/AtomicInteger.java
@@ -24,7 +24,7 @@ import java.io.IOException;
 public final class AtomicInteger extends java.util.concurrent.atomic.AtomicInteger implements RefI, X10JavaSerializable {
 
     private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(AtomicInteger.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(AtomicInteger.class);
 
     // constructor just for allocation
     public AtomicInteger(java.lang.System[] $dummy) {
@@ -74,7 +74,7 @@ public final class AtomicInteger extends java.util.concurrent.atomic.AtomicInteg
         return ai;
 	}
 
-	public int $_get_serialization_id() {
+	public short $_get_serialization_id() {
 		return _serialization_id;
 	}
 }
